fixed lives, randomised respawns

+++ helpers.h
@@ -9,7 +9,7 @@
 
 void check_pa_err(char *where, PaError err);
 PaStream * portaudio_init(PaStreamCallback *callback, float *data);
-float rand_between(int min, int max);
+float rand_between(float min, float max);
 void die(char *message, ...);
 void free_resources(ALLEGRO_DISPLAY *display, object *objects, ALLEGRO_EVENT_QUEUE *event_queue, PaStream *audio_stream);
 void load_respawn(object *objects, int n);
